refactor: really fix adhoc creation
