+++ src/ape_geth/query.py
@@ -4,7 +4,7 @@ from typing import Iterator, Optional
 from ape.api import ReceiptAPI
 from ape.api.query import ContractCreationQuery, QueryAPI, QueryType
 from ape.exceptions import QueryEngineError
-from ape_geth.provider import BaseGethProvider
+from ape_ethereum.provider import EthereumNodeProvider
 
 
 class OTSQueryEngine(QueryAPI):
@@ -21,7 +21,7 @@ class OTSQueryEngine(QueryAPI):
     @estimate_query.register
     def estimate_contract_creation_query(self, query: ContractCreationQuery) -> Optional[int]:
         if provider := self.network_manager.active_provider:
-            if not isinstance(provider, BaseGethProvider):
+            if not isinstance(provider, EthereumNodeProvider):
                 return None
             elif uri := provider.http_uri:
                 return 225 if uri.startswith("http://") else 600
@@ -30,6 +30,6 @@ class OTSQueryEngine(QueryAPI):
 
     @perform_query.register
     def get_contract_creation_receipt(self, query: ContractCreationQuery) -> Iterator[ReceiptAPI]:
-        if self.network_manager.active_provider and isinstance(self.provider, BaseGethProvider):
+        if self.network_manager.active_provider and isinstance(self.provider, EthereumNodeProvider):
             if receipt := self.provider._get_contract_creation_receipt(query.contract):
                 yield receipt
